Added solution for Part 4c and 4d (Ex 4.15-4.23)

+++ part4/bloglist/controllers/bloglist.js
@@ -1,8 +1,11 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 } )
   response.json(blogs)
 })
 
@@ -17,11 +20,14 @@ blogRouter.get('/:id', async (request, response, next) => {
 
 blogRouter.post('/', async (request, response, next) => {
   const body = request.body
+  const user = request.user
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user._id
   })
 
   if (blog.likes === undefined) {
@@ -29,11 +35,26 @@ blogRouter.post('/', async (request, response, next) => {
   }
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.json(savedBlog)
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (!user) {
+    return response.status(401).json({ error: 'user not found' })
+  } else if (!blog) {
+    return response.status(401).json({ error: 'blog not found' })
+  } else if (!blog.user) {
+    return response.status(401).json({ error: 'invalid blog user' })
+  } else if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+  }
+
   response.status(204).end()
 })
